Moved plotting functions to icaro folder
